Fire the env-mode lock check before conditional-build
vite-serve.js's first step is `execFileSync` of boxel-ui's
conditional-build.sh, which walks the addon's src+dist trees and may
trigger a full `pnpm run build` of @cardstack/boxel-ui — multi-second
work that runs before vite-with-traefik is even required. So a
second `BOXEL_ENVIRONMENT=… pnpm --filter @cardstack/host start`
sat silently for a while before the lock-refusal message appeared,
looking like a hang.

Extract the lockfile helpers into packages/host/scripts/env-mode-lock.js
and call `refuseIfAnotherSlugLocked()` from vite-serve.js and
serve-dist.js as the very first statement, before any other require
or exec. A spawned launcher now exits with the refuse message in
~30ms (measured) instead of waiting on conditional-build. The check
inside startWithTraefik stays as belt-and-suspenders for direct
callers of that function.
