+++ packages/host/scripts/vite-with-traefik.js
@@ -20,96 +20,12 @@
 require('./wtfnode-on-signal');
 const { spawn } = require('child_process');
 const path = require('path');
-const fs = require('fs');
 const net = require('net');
-
-// Embroider writes node_modules/.embroider/content-for.json once at boot
-// with the URL-encoded `config/environment.js` ENV blob, then serves it
-// to every `index.html` request. Two env-mode vites launched from the
-// same packages/host share that file: whichever booted last wins, and
-// `host.<slug-1>.localhost` ends up serving `<slug-2>`'s URLs even
-// though Traefik is routing to the correct port. Use a small lockfile
-// next to the cache to refuse the second startup with a clear message.
-// One env per worktree is the supported pattern.
-const ENV_MODE_LOCK_PATH = path.resolve(
-  __dirname,
-  '..',
-  'node_modules',
-  '.embroider',
-  '.env-mode-lock',
-);
-
-function isPidAlive(pid) {
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch (e) {
-    // EPERM means the process exists but is owned by someone else —
-    // still "alive" for our purposes (the embroider cache fight applies
-    // regardless of owner).
-    return e.code === 'EPERM';
-  }
-}
-
-function readEnvModeLock() {
-  let content;
-  try {
-    content = fs.readFileSync(ENV_MODE_LOCK_PATH, 'utf-8');
-  } catch (e) {
-    if (e.code !== 'ENOENT') {
-      console.warn(
-        `[environment-mode] Could not read ${ENV_MODE_LOCK_PATH}: ${e.message}`,
-      );
-    }
-    return null;
-  }
-  let [pidStr, lockedSlug] = content.trim().split(/\s+/, 2);
-  let pid = Number(pidStr);
-  if (!pid || !lockedSlug) return null;
-  return { pid, slug: lockedSlug };
-}
-
-function refuseIfAnotherSlugLocked(currentSlug) {
-  let lock = readEnvModeLock();
-  if (!lock) return;
-  if (lock.slug === currentSlug) return; // same env, idempotent
-  if (!isPidAlive(lock.pid)) return; // stale, will be overwritten below
-  console.error(
-    '\n[environment-mode] Refusing to start: another env-mode vite is already\n' +
-      `running from this worktree's packages/host (PID ${lock.pid},\n` +
-      `BOXEL_ENVIRONMENT slug "${lock.slug}"). This worktree's embroider cache\n` +
-      `at packages/host/node_modules/.embroider/ is shared across processes,\n` +
-      'so a second env launched here would silently make both vites serve\n' +
-      'identical HTML for whichever started last (the upstream port routing\n' +
-      "is fine — it's the bundled config/environment that gets clobbered).\n" +
-      '\n' +
-      'Use a separate git worktree per environment; see the "Environment\n' +
-      'mode: parallel environments" section of the repo-root README.\n',
-  );
-  process.exit(1);
-}
-
-function writeEnvModeLock(slug) {
-  try {
-    fs.mkdirSync(path.dirname(ENV_MODE_LOCK_PATH), { recursive: true });
-    fs.writeFileSync(ENV_MODE_LOCK_PATH, `${process.pid} ${slug}\n`, 'utf-8');
-  } catch (e) {
-    console.warn(
-      `[environment-mode] Could not write ${ENV_MODE_LOCK_PATH}: ${e.message}`,
-    );
-  }
-}
-
-function removeEnvModeLockIfOwned() {
-  let lock = readEnvModeLock();
-  if (lock && lock.pid === process.pid) {
-    try {
-      fs.unlinkSync(ENV_MODE_LOCK_PATH);
-    } catch {
-      /* already gone */
-    }
-  }
-}
+const {
+  refuseIfAnotherSlugLocked,
+  writeEnvModeLock,
+  removeEnvModeLockIfOwned,
+} = require('./env-mode-lock');
 
 function runVite({ subcommand, port, allHosts, host, extraEnv, nodeMemory }) {
   const args = ['vite'];
@@ -377,7 +293,10 @@ function startWithTraefik({ subcommand, defaultPort, label, nodeMemory }) {
   const { getEnvSlug, registerWithTraefik } = require('./traefik-helpers');
 
   const slug = getEnvSlug();
-  refuseIfAnotherSlugLocked(slug);
+  // Belt-and-suspenders: vite-serve.js / serve-dist.js also call this
+  // up-front (before boxel-ui conditional-build) so the second start
+  // exits immediately. Repeat here for direct callers of startWithTraefik.
+  refuseIfAnotherSlugLocked();
 
   ensureTraefik();
